% Auto-generated: do not edit by hand
\name{'dam'Footer}

\alias{'dam'Footer}

\title{Footer component}

\description{
Footer displays a resizable chat interface with scroll tracking and message submission.  @component @param {Object} props - Component props. @returns {JSX.Element} Rendered Footer component.
}

\usage{
'dam'Footer(children=NULL, id=NULL, message=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Child component to display in the footer.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{message}{Character. Message submitted by the user.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

