% Auto-generated: do not edit by hand
\name{'dam'Toggle}

\alias{'dam'Toggle}

\title{Toggle component}

\description{
ChatWindow displays a resizable chat interface with scroll tracking and message submission.  @component @param {Object} props - Component props. @param {Function} props.setProps - Dash callback for updating props. @param {boolean} props.isVisible - Indicates if the chat window is visible. @param {Function} props.children - Child component to display when the toggle is active. @returns {JSX.Element} Rendered ChatWindow component.
}

\usage{
'dam'Toggle(children=NULL, id=NULL, isVisible=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Child component to display when the toggle is active.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{isVisible}{Logical. Indicates of a toggle is visible (True) or closed (False).}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

